Add updateLocations() method to AirTracker class

+++ app/src/main/java/com/example/blank/AircraftDataStructure.java
@@ -123,7 +123,7 @@ public class AircraftDataStructure {
         return inWindow;
     }
 
-    public void updateLocations(float posLon, float posLat, float posAlt) {
+    public void updateLocations(double posLon, double posLat, double posAlt) {
         for (int azIndex = 0; azIndex < ARRAY_LENGTH; azIndex++) {
             for (int pitIndex = 0; pitIndex < ARRAY_LENGTH; pitIndex++) {
                 for (int i = 0; i < this.mAircraftBuckets[azIndex][pitIndex].size(); i++) {
